split layers module test in two.
